+++ Src/Private/Add-DiaHtmlNodeTable.ps1
@@ -122,6 +122,9 @@ Function Add-DiaHTMLNodeTable {
 
     #>
 
+    [CmdletBinding()]
+    [Alias('Get-DiaHTMLNodeTable')]
+    [OutputType([System.String])]
     param(
         [Parameter(
             Mandatory = $true,