+++ Tests/Get-NodeIP.Tests.ps1
@@ -7,4 +7,7 @@ Describe Get-NodeIP {
     It "Should return Host IP Address" {
         Get-NodeIP -Hostname localhost | Should -Be '127.0.0.1'
     }
+    It "Should return Unknown" {
+        Get-NodeIP -Hostname 'invalid-hostname'  | Should -Be 'Unknown'
+    }
 }